Refactoring: Move RateLimiting into PaginatedRequestFetcher
